fourth attempt to fix up y orientation

+++ Geometry/CommonTopologies/src/RadialStripTopology.cc
@@ -28,7 +28,7 @@ RadialStripTopology::RadialStripTopology(int ns, float aw, float dh, float r, in
 float 
 RadialStripTopology::xOfStrip(int strip, float y) const {
   // Expect input 'strip' to be in range 1 to nstrips()
-  float tanPhi = tan( stripAngle(static_cast<float>(strip) - 0.5 ) );
+  double tanPhi = tan( stripAngle(static_cast<float>(strip) - 0.5 ) );
   return yAxisOrientation()*( y*yAxisOrientation()+originToIntersection() ) * tanPhi;
 }
 
@@ -39,15 +39,15 @@ RadialStripTopology::localPosition(float strip) const {
 
 LocalPoint 
 RadialStripTopology::localPosition(const MeasurementPoint& mp) const {
-  float phi = stripAngle( mp.x() );
+  double phi = stripAngle( mp.x() );
   // 2nd component of MP is fractional position along strip, range +/-0.5,
   // so distance along strip (measured from mid-point of length of strip) is
   //     mp.y() * (length of strip). 
   // Thus distance along y direction, from mid-point of strip, is 
   //    mp.y() * (length of strip) * cos(phi)
   // But (length of strip) = theDetHeight/cos(phi), so
-  float y =  mp.y() * theDetHeight + yCentreOfStripPlane();
-  float x = yAxisOrientation() * ( originToIntersection() + y * yAxisOrientation() ) * tan ( phi );
+  double y =  mp.y() * theDetHeight + yCentreOfStripPlane();
+  double x = yAxisOrientation() * ( originToIntersection() + y * yAxisOrientation() ) * tan ( phi );
   return LocalPoint( x, y );
 }
 
@@ -57,24 +57,24 @@ RadialStripTopology::localError(float strip, float stripErr2) const {
   // Since 'strip' is in units of angular strip-widths, stripErr2 is
   // required to be in corresponding units.
 
-  float t = tan( stripAngle( strip ) );        // tan(angle between strip and y)
-  float c2 = 1./(1. + t*t);                    // cos(angle)**2
-  float cs = t*c2;                             // sin(angle)*cos(angle); tan carries sign of sin!
-  float s2 = t*t * c2;                         // sin(angle)**2
+  double t = tan( stripAngle( strip ) );        // tan(angle between strip and y)
+  double c2 = 1./(1. + t*t);                    // cos(angle)**2
+  double cs = t*c2;                             // sin(angle)*cos(angle); tan carries sign of sin!
+  double s2 = t*t * c2;                         // sin(angle)**2
 
-  float D2 = originToIntersection()*originToIntersection() / c2;
-  float L2 = theDetHeight*theDetHeight / c2;   // length**2 of strip across detector
-  float A2 = theAngularWidth*theAngularWidth;
+  double D2 = originToIntersection()*originToIntersection() / c2;
+  double L2 = theDetHeight*theDetHeight / c2;   // length**2 of strip across detector
+  double A2 = theAngularWidth*theAngularWidth;
 
   // The error**2 we're assigning is L2/12 wherever we measure the position along the strip
   // from... we just know the measurement is somewhere ON the strip. 
 
-  float SD2 = L2 / 12.;       // SR = Sigma-Radius ('Radius' is along strip)
-  float SA2 = A2 * stripErr2; // SA = Sigma-Angle
+  double SD2 = L2 / 12.;       // SR = Sigma-Radius ('Radius' is along strip)
+  double SA2 = A2 * stripErr2; // SA = Sigma-Angle
 
-  float sx2 = c2 * D2 * SA2 + s2 * SD2;
-  float sy2 = s2 * D2 * SA2 + c2 * SD2;
-  float rhosxsy = cs * ( SD2 - D2 * SA2 );
+  double sx2 = c2 * D2 * SA2 + s2 * SD2;
+  double sy2 = s2 * D2 * SA2 + c2 * SD2;
+  double rhosxsy = cs * ( SD2 - D2 * SA2 );
 
   return LocalError(sx2, rhosxsy, sy2);
 }
@@ -85,35 +85,35 @@ RadialStripTopology::localError(const MeasurementPoint& mp,
   // Here we need to allow the possibility of correlated errors, since
   // that may happen during Kalman filtering
 
-  float phi = thePhiOfOneEdge + mp.x() * theAngularWidth;
+  double phi = thePhiOfOneEdge + mp.x() * theAngularWidth;
 
-  float t = tan( phi );                        // tan(angle between strip and y)
-  float c2 = 1./(1. + t*t);                    // cos(angle)**2
-  float cs = t*c2;                             // sin(angle)*cos(angle); tan carries sign of sin!
-  float s2 = t*t * c2;                         // sin(angle)**2
+  double t = tan( phi );                        // tan(angle between strip and y)
+  double c2 = 1./(1. + t*t);                    // cos(angle)**2
+  double cs = t*c2;                             // sin(angle)*cos(angle); tan carries sign of sin!
+  double s2 = t*t * c2;                         // sin(angle)**2
 
-  float A  = theAngularWidth;
-  float A2 = A * A;
+  double A  = theAngularWidth;
+  double A2 = A * A;
 
   // D is distance from intersection of edges to hit on strip
-  float D = (originToIntersection() + mp.y() * yAxisOrientation() * theDetHeight) / sqrt(c2);
-  float D2 = D * D;
+  double D = (originToIntersection() + mp.y() * yAxisOrientation() * theDetHeight) / sqrt(c2);
+  double D2 = D * D;
 
   // L is length of strip across face of chamber
-  float L2 = theDetHeight*theDetHeight / c2;  
-  float L  = sqrt(L2); 
+  double L2 = theDetHeight*theDetHeight / c2;  
+  double L  = sqrt(L2); 
 
   // MeasurementError elements are already squared
   // but they're normalized to products of A and L (N.B. L not D!)
   // @@ ENSURE MEASUREMENT ERROR COMPONENTS ARE INDEED NORMALIZED LIKE THIS!
 
-  float SA2 = merr.uu() * A2;
-  float SD2 = merr.vv() * L2; // Note this norm uses stripLength**2
-  float RHOSASR = merr.uv() * A * L;
+  double SA2 = merr.uu() * A2;
+  double SD2 = merr.vv() * L2; // Note this norm uses stripLength**2
+  double RHOSASR = merr.uv() * A * L;
 
-  float sx2 = SA2 * D2 * c2  +  2. * RHOSASR * D * cs  +  SD2 * s2;
-  float sy2 = SA2 * D2 * s2  -  2. * RHOSASR * D * cs  +  SD2 * c2;
-  float rhosxsy = cs * ( SD2 - D2 * SA2 )  +  RHOSASR * D * ( c2 - s2 );
+  double sx2 = SA2 * D2 * c2  +  2. * RHOSASR * D * cs  +  SD2 * s2;
+  double sy2 = SA2 * D2 * s2  -  2. * RHOSASR * D * cs  +  SD2 * c2;
+  double rhosxsy = cs * ( SD2 - D2 * SA2 )  +  RHOSASR * D * ( c2 - s2 );
 
   return LocalError(sx2, rhosxsy, sy2);
 }
@@ -122,8 +122,8 @@ float
 RadialStripTopology::strip(const LocalPoint& lp) const {
   // Note that this phi is (pi/2 - conventional local phi)
   // This means use atan2(x,y) rather than more usual atan2(y,x)
-  float phi = yAxisOrientation() * atan2( lp.x(), lp.y()*yAxisOrientation()+originToIntersection() );
-  float aStrip = yAxisOrientation() * (phi-thePhiOfOneEdge)/theAngularWidth;
+  double phi = yAxisOrientation() * atan2( lp.x(), lp.y()*yAxisOrientation()+originToIntersection() );
+  double aStrip = yAxisOrientation() * (phi-thePhiOfOneEdge)/theAngularWidth;
   aStrip = (aStrip >= 0. ? aStrip : 0.);
   aStrip = (aStrip <= theNumberOfStrips ? aStrip : theNumberOfStrips);
   return aStrip;
@@ -133,7 +133,7 @@ MeasurementPoint
 RadialStripTopology::measurementPosition(const LocalPoint& lp) const {
   // Note that this phi is (pi/2 - conventional local phi)
   // This means use atan2(x,y) rather than more usual atan2(y,x)
-  float phi = yAxisOrientation() * atan2( lp.x(), lp.y()*yAxisOrientation()+originToIntersection() );
+  double phi = yAxisOrientation() * atan2( lp.x(), lp.y()*yAxisOrientation()+originToIntersection() );
   return MeasurementPoint( yAxisOrientation()*(phi-thePhiOfOneEdge)/theAngularWidth,
 			   lp.y()/theDetHeight );
 }
@@ -142,29 +142,29 @@ MeasurementError
 RadialStripTopology::measurementError(const LocalPoint& lp,
 				      const LocalError& lerr) const {
 
-  float yHitToInter = lp.y()*yAxisOrientation() + originToIntersection();
+  double yHitToInter = lp.y()*yAxisOrientation() + originToIntersection();
   // Care! sign of angle measurement must be consistently treated when yAxis orientation changes.
-  float t  = yAxisOrientation() * lp.x() / yHitToInter; // tan(angle between strip and y) 
-  float c2 = 1./(1. + t*t);  // cos(angle)**2
-  float cs = t*c2;           // sin(angle)*cos(angle); tan carries sign of sin!
-  float s2 = t*t * c2;       // sin(angle)**2
+  double t  = yAxisOrientation() * lp.x() / yHitToInter; // tan(angle between strip and y) 
+  double c2 = 1./(1. + t*t);  // cos(angle)**2
+  double cs = t*c2;           // sin(angle)*cos(angle); tan carries sign of sin!
+  double s2 = t*t * c2;       // sin(angle)**2
 
-  float A  = theAngularWidth;
+  double A  = theAngularWidth;
 
   // L is length of strip across face of chamber
-  float L2 = theDetHeight*theDetHeight / c2;  
-  float L  = sqrt(L2); 
+  double L2 = theDetHeight*theDetHeight / c2;  
+  double L  = sqrt(L2); 
 
   // D is distance from intersection of edges to hit on strip
-  float D2 = lp.x()*lp.x() + yHitToInter*yHitToInter;
-  float D = sqrt(D2);
+  double D2 = lp.x()*lp.x() + yHitToInter*yHitToInter;
+  double D = sqrt(D2);
 
-  float LP = D * A;
-  float LP2 = LP * LP;
+  double LP = D * A;
+  double LP2 = LP * LP;
 
-  float SA2 = ( c2 * lerr.xx() - 2. * cs * lerr.xy() + s2 * lerr.yy() )/LP2;
-  float SD2 =   s2 * lerr.xx() + 2. * cs * lerr.xy() + c2 * lerr.yy()/L2;
-  float RHOSASR = ( cs * ( lerr.xx() - lerr.yy() ) + ( c2 - s2 ) * lerr.xy() )/LP/L;
+  double SA2 = ( c2 * lerr.xx() - 2. * cs * lerr.xy() + s2 * lerr.yy() ) / LP2;
+  double SD2 = ( s2 * lerr.xx() + 2. * cs * lerr.xy() + c2 * lerr.yy() ) / L2;
+  double RHOSASR = ( cs * ( lerr.xx() - lerr.yy() ) + ( c2 - s2 ) * lerr.xy() ) / LP / L;
 
   return MeasurementError(SA2, RHOSASR, SD2);
 }
@@ -191,11 +191,11 @@ RadialStripTopology::localPitch(const LocalPoint& lp) const {
   // Calculating it is a nightmare...deriving the expression below is left
   // as a exercise for the reader. 
 
-  float fstrip = strip(lp); // position in strip units
+  double fstrip = strip(lp); // position in strip units
   int istrip = static_cast<int>(fstrip + 1.0); // which strip number
   istrip = (istrip>nstrips() ? nstrips() : istrip); // enforce maximum
-  float fangle = stripAngle(static_cast<float>(istrip - 0.5)); // angle of strip centre
-  float localp = ( lp.y()*yAxisOrientation() + originToIntersection() ) * sin(theAngularWidth) /
+  double fangle = stripAngle(static_cast<float>(istrip - 0.5)); // angle of strip centre
+  double localp = ( lp.y()*yAxisOrientation() + originToIntersection() ) * sin(theAngularWidth) /
     ( cos(fangle-theAngularWidth/2.)*cos(fangle+theAngularWidth/2.) );
   return localp;
 }
@@ -206,10 +206,10 @@ RadialStripTopology::stripAngle(float strip) const {
 }
   
 float RadialStripTopology::localStripLength(const LocalPoint& lp) const {
-  float yHitToInter = lp.y()*yAxisOrientation() + originToIntersection();
+  double yHitToInter = lp.y()*yAxisOrientation() + originToIntersection();
   // since we're dealing with magnitudes, sign is unimportant
-  float t  = lp.x() / yHitToInter;    // tan(angle between strip and y)
-  float c2 = 1./(1. + t*t);           // cos(angle)**2
+  double t  = lp.x() / yHitToInter;    // tan(angle between strip and y)
+  double c2 = 1./(1. + t*t);           // cos(angle)**2
   return theDetHeight / sqrt(c2);
 }
 
@@ -218,7 +218,7 @@ int RadialStripTopology::nearestStrip(const LocalPoint & lp) const
   // xxxStripTopology::strip() is expected to have range 0. to 
   // float(no of strips), but be extra careful and enforce that
 
-  float fstrip = this->strip(lp);
+  double fstrip = this->strip(lp);
   int n_strips = this->nstrips();
   fstrip = ( fstrip>=0. ? fstrip :  0. ); // enforce minimum 0.
   int near = static_cast<int>( fstrip ) + 1; // first strip is 1
